// types/preference.ts
export type Preference =
  | 'streetFood'
  | 'bento'
  | 'restaurant'
  | 'spicy'
  | 'seafood'
  | 'vegetarian'
  | 'lowCarbs'
  | 'light'
  | 'strong'
  | 'meat'
  | 'vegetables'
  | 'protein'
  | 'lowFat'
  | 'lowCalories'
  | 'rice'
  | 'noodles'
  | 'saucy'
  | 'dessert'
  | 'snack'
  | 'chinese'
  | 'american'
  | 'european'
  | 'mediterranean'
  | 'middleEastern'
  | 'indian'
  | 'japanese'
  | 'korean'
  | 'southeastAsian';